WIP: Rearranging implementation files, API work
