+++ .gitignore
@@ -265,3 +265,4 @@ Desktop.ini
 $RECYCLE.BIN/
 logo.h
 __pycache__
+core/method_bind_ext.inc
