+++ core/io/formats/file_access_hpck.h
@@ -1,0 +1,11 @@
+#ifndef FILE_ACCESS_HPCK_H
+#define FILE_ACCESS_HPCK_H
+
+#include "core/io/file_access_pack.h"
+
+//This is an implementation of pck that only uses hashes for files, which
+//reduces file size of the end archive
+
+
+
+#endif
